Possible collaboration for new web browser (Xcode 6)
Hello! I have no experience with coding, GitHub and developing. I made
this little web browser called Lite Browser in XCode 6 for OS X
Yosemite. The browser is pretty useless at its current state, however,
I was wondering if anybody would be interested in helping me develop
this. I don’t expect Lite Browser to come close to being as popular as
Chrome or Firefox, but I want to create a nice, stable browser. The
name Lite Browser is going to change in the future, but for now, it’s
nice and simple. Thanks for reading and contact me if you’re interested
in helping me with this project!